requirements.txt
django
ffmpeg-python
django-crispy-forms
crispy-bootstrap5